refactor: enums package
