+++ src/model/Task.java
@@ -1,7 +1,7 @@
 package model;
 
-import util.Status;
-import util.Type;
+import util.enums.Status;
+import util.enums.Type;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
