+++ src/util/TaskTimeController.java
@@ -3,6 +3,7 @@ package util;
 import model.Epic;
 import model.SubTask;
 import model.Task;
+import util.enums.Type;
 
 import java.time.LocalDateTime;
 import java.util.Comparator;
