+++ src/util/enums/CsvField.java
@@ -1,4 +1,4 @@
-package util;
+package util.enums;
 
 public enum CsvField {
     ID(0),
