+++ src/util/enums/Status.java
@@ -1,4 +1,4 @@
-package util;
+package util.enums;
 
 public enum Status {
     NEW,
