+++ test/managers/InMemoryHistoryManagerTest.java
@@ -5,7 +5,7 @@ import model.Epic;
 import model.SubTask;
 import model.Task;
 import org.junit.jupiter.api.Test;
-import util.Status;
+import util.enums.Status;
 
 import static org.junit.jupiter.api.Assertions.*;
