+++ test/model/EpicTest.java
@@ -1,7 +1,7 @@
 package model;
 
 import org.junit.jupiter.api.Test;
-import util.Status;
+import util.enums.Status;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
